[PJ-42] [ALL] API quote subscription symbol limit to only N(default to 5)
Limit the number of symbols that API can subscribe to only N(default to 5). If user subscribe to
6th symbol, the first one will be cancelled internally such that the total number is still 5.
Also include code refactoring described below:
The original data type of quote subscription is Map<String, Map<String, String>>
which is composed of "symbol", "user", "symbol" and where outer key is the same as inner value.
Thus refactor it to be Map<String, List<String>> and key is "user", values are list of "symbol".
Related logic flow is also updated accordingly.